Make MCL_STATIC_LIB full static
Currently MCL_STATIC_LIB will produce static mcl::mclbn{N} targets, but
these static targets still link to the *non*-static mcl::mcl and so you
end up with a half-static result that still depends on a dynamic lib.

This fixes it to link to the mcl_st static target when MCL_STATIC_LIB is
enabled.